+++ packages/mars-core/src/h5/index.js
@@ -1,7 +1,7 @@
 /**
  * @file  h5 runtime entry
+ * @description this file is for H5 runtime, and will be copied directly without compile or pack
  * @author zhangwentao
  */
-
 export {default as Vue} from 'vue';
-export {default as config} from '../config';
+export const config = {};