+++ packages/mars-cli-template/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.14](https://github.com/max-team/Mars/compare/@marsjs/cli-template@0.1.13...@marsjs/cli-template@0.1.14) (2019-07-19)
+
+
+### Features
+
+* **cli-template:** add block config and PascalCase example ([91209c7](https://github.com/max-team/Mars/commit/91209c7))
+
+
+
+
+
 ## [0.1.13](https://github.com/max-team/Mars/compare/@marsjs/cli-template@0.1.12...@marsjs/cli-template@0.1.13) (2019-07-11)
 
 **Note:** Version bump only for package @marsjs/cli-template